added recommendations, fixed revenue overflow

+++ app/src/main/java/com/projetb3/movynov/dataclasses/MediaMovie.kt
@@ -29,7 +29,7 @@ class MediaMovie(
     @SerializedName("popularity"            ) var popularity          : Double?                        = null,
     @SerializedName("poster_path"           ) var posterPath          : String?                        = null,
     @SerializedName("release_date"          ) var releaseDate         : String?                        = null,
-    @SerializedName("revenue"               ) var revenue             : Int?                           = null,
+    @SerializedName("revenue"               ) var revenue             : Long?                           = null,
     @SerializedName("runtime"               ) var runtime             : Int?                           = null,
     @SerializedName("status"                ) var status              : String?                        = null,
     @SerializedName("tagline"               ) var tagline             : String?                        = null,
